# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * GradHub API – student project & teammate finder for Ain Shams Engineering
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ApiError {
  error: string;
}

export interface MessageResponse {
  message: string;
}

export interface SignupInput {
  /** @minLength 2 */
  name: string;
  /** Must end with @eng.asu.edu.eg */
  email: string;
  /** @minLength 8 */
  password: string;
  major: string;
  skills: string[];
  bio?: string;
  /** @nullable */
  phone?: string | null;
}

export interface LoginInput {
  email: string;
  password: string;
}

export type ProfileViewRole = typeof ProfileViewRole[keyof typeof ProfileViewRole];


export const ProfileViewRole = {
  student: 'student',
  admin: 'admin',
} as const;

/**
 * Connect status relative to the authenticated viewer
 */
export type ProfileViewConnectStatus = typeof ProfileViewConnectStatus[keyof typeof ProfileViewConnectStatus];


export const ProfileViewConnectStatus = {
  none: 'none',
  pending_sent: 'pending_sent',
  pending_received: 'pending_received',
  connected: 'connected',
} as const;

/**
 * Public-safe user view – never includes email or phone
 */
export interface ProfileView {
  id: number;
  name: string;
  major: string;
  skills: string[];
  /** @nullable */
  bio?: string | null;
  role: ProfileViewRole;
  createdAt: string;
  /** Connect status relative to the authenticated viewer */
  connectStatus: ProfileViewConnectStatus;
}

export type UserDetailViewRole = typeof UserDetailViewRole[keyof typeof UserDetailViewRole];


export const UserDetailViewRole = {
  student: 'student',
  admin: 'admin',
} as const;

export type UserDetailViewConnectStatus = typeof UserDetailViewConnectStatus[keyof typeof UserDetailViewConnectStatus];


export const UserDetailViewConnectStatus = {
  none: 'none',
  pending_sent: 'pending_sent',
  pending_received: 'pending_received',
  connected: 'connected',
} as const;

/**
 * Profile view plus optional contact info when connected
 */
export interface UserDetailView {
  id: number;
  name: string;
  major: string;
  skills: string[];
  /** @nullable */
  bio?: string | null;
  role: UserDetailViewRole;
  createdAt: string;
  connectStatus: UserDetailViewConnectStatus;
  /**
     * Only present when connected
     * @nullable
     */
  email?: string | null;
  /**
     * Only present when connected and phone is set
     * @nullable
     */
  phone?: string | null;
}

export interface ProfileUpdate {
  name?: string;
  major?: string;
  skills?: string[];
  /** @nullable */
  bio?: string | null;
  /** @nullable */
  phone?: string | null;
}

export type ProjectStatus = typeof ProjectStatus[keyof typeof ProjectStatus];


export const ProjectStatus = {
  open: 'open',
  in_progress: 'in_progress',
  closed: 'closed',
} as const;

export interface Project {
  id: number;
  title: string;
  description: string;
  requiredSkills: string[];
  status: ProjectStatus;
  ownerId: number;
  ownerName: string;
  teamSizeCap: number;
  createdAt: string;
}

export type ProjectDetailStatus = typeof ProjectDetailStatus[keyof typeof ProjectDetailStatus];


export const ProjectDetailStatus = {
  open: 'open',
  in_progress: 'in_progress',
  closed: 'closed',
} as const;

/**
 * Viewer's connect status with the project owner
 */
export type ProjectDetailConnectStatus = typeof ProjectDetailConnectStatus[keyof typeof ProjectDetailConnectStatus];


export const ProjectDetailConnectStatus = {
  none: 'none',
  pending_sent: 'pending_sent',
  pending_received: 'pending_received',
  connected: 'connected',
} as const;

export interface ProjectDetail {
  id: number;
  title: string;
  description: string;
  requiredSkills: string[];
  status: ProjectDetailStatus;
  ownerId: number;
  ownerName: string;
  teamSizeCap: number;
  createdAt: string;
  /** Viewer's connect status with the project owner */
  connectStatus: ProjectDetailConnectStatus;
}

export type ProjectInputStatus = typeof ProjectInputStatus[keyof typeof ProjectInputStatus];


export const ProjectInputStatus = {
  open: 'open',
  in_progress: 'in_progress',
  closed: 'closed',
} as const;

export interface ProjectInput {
  /** @minLength 3 */
  title: string;
  /** @minLength 10 */
  description: string;
  requiredSkills: string[];
  status?: ProjectInputStatus;
  /**
     * @minimum 1
     * @maximum 20
     */
  teamSizeCap: number;
}

export type ProjectUpdateStatus = typeof ProjectUpdateStatus[keyof typeof ProjectUpdateStatus];


export const ProjectUpdateStatus = {
  open: 'open',
  in_progress: 'in_progress',
  closed: 'closed',
} as const;

export interface ProjectUpdate {
  title?: string;
  description?: string;
  requiredSkills?: string[];
  status?: ProjectUpdateStatus;
  teamSizeCap?: number;
}

export type ConnectRequestStatus = typeof ConnectRequestStatus[keyof typeof ConnectRequestStatus];


export const ConnectRequestStatus = {
  pending: 'pending',
  accepted: 'accepted',
  declined: 'declined',
} as const;

export interface ConnectRequest {
  id: number;
  senderId: number;
  senderName: string;
  recipientId: number;
  recipientName: string;
  /** @nullable */
  projectId?: number | null;
  /** @nullable */
  projectTitle?: string | null;
  status: ConnectRequestStatus;
  createdAt: string;
}

export interface ConnectionList {
  incoming: ConnectRequest[];
  outgoing: ConnectRequest[];
}

export interface ConnectionInput {
  recipientId: number;
  /** @nullable */
  projectId?: number | null;
}

export type ConnectionResponseStatus = typeof ConnectionResponseStatus[keyof typeof ConnectionResponseStatus];


export const ConnectionResponseStatus = {
  accepted: 'accepted',
  declined: 'declined',
} as const;

export interface ConnectionResponse {
  status: ConnectionResponseStatus;
}

export interface Notification {
  id: number;
  userId: number;
  message: string;
  read: boolean;
  createdAt: string;
}

export interface DashboardStats {
  openProjects: number;
  totalStudents: number;
  myProjects: number;
  pendingRequests: number;
  unreadNotifications: number;
}

export type ListUsersParams = {
/**
 * Comma-separated skill tags to filter by
 */
skills?: string;
/**
 * Major to filter by
 */
major?: string;
/**
 * Free-text search across name and bio
 */
search?: string;
};

export type ListProjectsParams = {
/**
 * Comma-separated required-skill tags to filter by
 */
skills?: string;
/**
 * Filter by project status
 */
status?: ListProjectsStatus;
/**
 * Free-text search across title and description
 */
search?: string;
};

export type ListProjectsStatus = typeof ListProjectsStatus[keyof typeof ListProjectsStatus];


export const ListProjectsStatus = {
  open: 'open',
  in_progress: 'in_progress',
  closed: 'closed',
} as const;