# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: GradHub API – student project & teammate finder for Ain Shams Engineering
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /users:
    get:
      operationId: listUsers
      tags: [users]
      summary: List student profiles (public view — no contact info)
      parameters:
        - in: query
          name: skills
          schema:
            type: string
          description: Comma-separated skill tags to filter by
        - in: query
          name: major
          schema:
            type: string
          description: Major to filter by
        - in: query
          name: search
          schema:
            type: string
          description: Free-text search across name and bio
      responses:
        "200":
          description: List of profiles
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ProfileView"
  /users/{id}:
    get:
      operationId: getUser
      tags: [users]
      summary: Get a user profile (contact info included only if connected)
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: User profile (contact conditionally included)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserDetailView"
        "404":
          description: User not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
  /users/{id}/profile:
    patch:
      operationId: updateProfile
      tags: [users]
      summary: Update own profile
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProfileUpdate"
      responses:
        "200":
          description: Updated profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProfileView"
        "403":
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"

  # ── Projects ──────────────────────────────────────────────────────────
...
components:
  schemas:
...
    ApiError:
      type: object
      properties:
        error:
          type: string
      required:
        - error
...
    ProfileView:
      type: object
      description: Public-safe user view – never includes email or phone
      required: [id, name, major, skills, role, createdAt, connectStatus]
      properties:
        id:
          type: integer
        name:
          type: string
        major:
          type: string
        skills:
          type: array
          items:
            type: string
        bio:
          type: ["string", "null"]
        role:
          type: string
          enum: [student, admin]
        createdAt:
          type: string
          format: date-time
        connectStatus:
          type: string
          enum: [none, pending_sent, pending_received, connected]
          description: Connect status relative to the authenticated viewer
    UserDetailView:
      type: object
      description: Profile view plus optional contact info when connected
      required: [id, name, major, skills, role, createdAt, connectStatus]
      properties:
        id:
          type: integer
        name:
          type: string
        major:
          type: string
        skills:
          type: array
          items:
            type: string
        bio:
          type: ["string", "null"]
        role:
          type: string
          enum: [student, admin]
        createdAt:
          type: string
          format: date-time
        connectStatus:
          type: string
          enum: [none, pending_sent, pending_received, connected]
        email:
          type: ["string", "null"]
          description: Only present when connected
        phone:
          type: ["string", "null"]
          description: Only present when connected and phone is set
    ProfileUpdate:
      type: object
      properties:
        name:
          type: string
        major:
          type: string
        skills:
          type: array
          items:
            type: string
        bio:
          type: ["string", "null"]
        phone:
          type: ["string", "null"]

    # ── Project ─────────────────────────────────────────────────────────
...